Import openssl-0.9.7a-blinding-threads.patch from dist-git
Try to fix thread-safety for RSA blinding without breaking the public ABI.  In
OpenSSL 0.9.7b, safety is checked by adding a thread ID field to the (public)
RSA blinding structure, comparing it to the current thread ID when performing
RSA encryption or decryption, and using a temporary blinding structure if they
don't match.  In this patch we attempt to do mostly the same thing, but instead
of storing the thread ID in the blinding structure, we store it as an ex_data
item in the RSA key structure itself.  We add a private function for checking
if we're the "owning" thread (OPENSSL_private_rsa_blinding_check_thread_id),
and wrap the creation of the ex_data item and the comparison inside of the
RSA_BLINDING lock.  Adding a private function doesn't require changing any of
the public headers, so the result remains binary-compatible with unpatched
OpenSSL 0.9.7a.  Of course, if you're moving from 0.9.7/0.9.7a to a newer
version of OpenSSL, you should just drop this patch.
 
commit 120e30eee3dd762d06b57da8f46f1686db822ecd
Author: cvsdist <>
Date:   Thu Sep 9 09:45:46 2004 +0000

    auto-import changelog data from openssl-0.9.7a-20.src.rpm
    * Wed Sep 24 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-20
    - only parse a client cert if one was requested
    - temporarily exclusivearch for %{ix86}
    
    * Tue Sep 23 2003 Nalin Dahyabhai <nalin@redhat.com>
    - add security fixes for protocol parsing bugs (CAN-2003-0543, CAN-2003-0544)
      and heap corruption (CAN-2003-0545)
    - update RHNS-CA-CERT files
    - ease back on the number of threads used in the threading test
    
    * Wed Sep 17 2003 Matt Wilson <msw@redhat.com> 0.9.7a-19
    - rebuild to fix gzipped file md5sums (#91211)
    
    * Mon Aug 25 2003 Phil Knirsch <pknirsch@redhat.com> 0.9.7a-18
    - Updated libica to version 1.3.4.
    
    * Thu Jul 17 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-17
    - rebuild
    
    * Tue Jul 15 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-10.9
    - free the kssl_ctx structure when we free an SSL structure (#99066)
    
    * Thu Jul 10 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-16
    - rebuild
    
    * Thu Jul 10 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-15
    - lower thread test count on s390x
    
    * Tue Jul 08 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-14
    - rebuild
    
    * Thu Jun 26 2003 Nalin Dahyabhai <nalin@redhat.com> 0.9.7a-13
    - disable assembly on arches where it seems to conflict with threading
    
    * Thu Jun 26 2003 Phil Knirsch <pknirsch@redhat.com> 0.9.7a-12
    - Updated libica to latest upstream version 1.3.0