Import openssl-0.9.7a-utf8fix.patch from dist-git
Somewhere deep inside of ASN1_STRING_to_UTF8() (used, among other things, to
parse PKCS12 certificates), the out_utf8() function is passed to a function
(traverse_string) which calls it iteratively.  In all cases where it is called,
arg points to an int, not a long, so we need to avoid screwing things up on
systems where sizeof(int) != sizeof(long).
 
commit f5e618ad460595db80ce9e6cf38c8d5af57822ab
Author: cvsdist <>
Date:   Thu Sep 9 09:47:19 2004 +0000

    auto-import openssl-0.9.7a-26  from openssl-0.9.7a-26.src.rpm

commit f240c3084e91e8c367489680c83a1c2f2f351f33
Author: cvsdist <>
Date:   Thu Sep 9 09:46:50 2004 +0000

    auto-import openssl-0.9.7a-23  from openssl-0.9.7a-23.src.rpm