Import openssl-0.9.7a-ssl-strict-matching.patch from dist-git
*) In the SSL/TLS server implementation, be strict about session ID
context matching (which matters if an application uses a single
external cache for different purposes).  Previously,
out-of-context reuse was forbidden only if SSL_VERIFY_PEER was
set. This did ensure strict client verification, but meant that,
with applications using a single external cache for quite
different requirements, clients could circumvent ciphersuite
restrictions for a given session ID context by starting a session
in a different context.
commit d2cc8c00ebce7ae9811e184d2c097d90401eb138
Author: Tomas Mraz <tmraz@redhat.com>
Date:   Thu Aug 2 20:33:13 2007 +0000

    * Thu Aug  2 2007 Tomas Mraz <tmraz@redhat.com> 0.9.7a-43.17
    - use poll when reading random device (#236164)
    - make ssl session ID context matching strict (#244436)
    - openssl utility shouldn't crash on invalid PKCS#12 files (#245083)
    - CVE-2007-3108 remove conditionals in BN_div, BN_mod and final
      Montgomery reduction (#250580)
    - Resolves: rhbz#236164 rhbz#244436 rhbz#245083 rhbz#250580