Import openssl-0.9.7a-ssl-strict-matching.patch from dist-git
*) In the SSL/TLS server implementation, be strict about session ID
context matching (which matters if an application uses a single
external cache for different purposes).  Previously,
out-of-context reuse was forbidden only if SSL_VERIFY_PEER was
set. This did ensure strict client verification, but meant that,
with applications using a single external cache for quite
different requirements, clients could circumvent ciphersuite
restrictions for a given session ID context by starting a session
in a different context.
commit 41baafcbcd3e0175c0fe67de33b3660ac9b391c5
Author: Tomas Mraz <tmraz@redhat.com>
Date:   Tue Jan 15 08:19:40 2008 +0000

    * Tue Jan 15 2008 Tomas Mraz <tmraz@redhat.com> 0.9.8b-10
    - compile with -march=z900 on s390 for performance improvements (#250818)
    - make ssl session ID matching strict (#233599)
    - Resolves: rhbz#250818 rhbz#233599