Import openssl-1.0.0-cve-2012-0884.patch from dist-git
Fix MMA (Bleichenbacher's attack on PKCS #1 v1.5 RSA padding) weakness
in CMS and PKCS7 code. When RSA decryption fails use a random key for
content decryption and always return the same error. Note: this attack
needs on average 2^20 messages so it only affects automated senders. The
old behaviour can be reenabled in the CMS code by setting the
CMS_DEBUG_DECRYPT flag: this is useful for debugging and testing where
an MMA defence is not necessary.
Thanks to Ivan Nestlerode <inestlerode@us.ibm.com> for discovering
this issue. (CVE-2012-0884)
[Steve Henson]
commit 96714380660fd05f54fd82fd884a3a53573bbe5a
Author: Tomas Mraz <tmraz@fedoraproject.org>
Date:   Tue May 15 11:48:41 2012 +0200

    Resolves: rhbz#821649
    
    fix for CVE-2012-2333 - improper checking for record length in DTLS (#820686)
    properly initialize tkeylen in the CVE-2012-0884 fix

commit 05b5c904e2320258c9aa64f9397be87bbb4b8f6e
Author: Tomas Mraz <tmraz@fedoraproject.org>
Date:   Mon Mar 19 17:11:40 2012 +0100

    Resolves: rhbz#804553
    
    - fix problem with the SGC restart patch that might terminate handshake
      incorrectly
    - fix for CVE-2012-0884 - MMA weakness in CMS and PKCS#7 code (#802725)
    - fix for CVE-2012-1165 - NULL read dereference on bad MIME headers (#802489)